Created PlayerSetup class to assign players
* Creates two types of Player objects based on the marker selected
