+++ lib/tic_tac_toe.rb
@@ -5,13 +5,13 @@ require_relative "game"
 require_relative "wrapper_io"
 require_relative "validation"
 require_relative "marker_loop"
-require_relative "game_turn"
 require_relative "game_completion"
 require_relative "player"
 require_relative "console_runner"
 require_relative "game_type"
 require_relative "computer"
 require_relative "human"
+require_relative "player_setup"
 
 module TicTacToe
 end
