+++ spec/game_type.rb
