+++ spec/player_setup_spec.rb
@@ -1,0 +1,26 @@
+module TicTacToe
+  require 'spec_helper'
+  describe PlayerSetup do
+
+    let(:mock_console_ui) { double }
+    let(:player_setup) { double }
+
+    context "human v. human" do 
+      it "returns the corresponding Player objects for each player" do
+        args = 
+          {
+            :board => Board.new, 
+            :console_ui => mock_console_ui,
+            :validation => Validation.new
+          }
+        marker1, marker2 = "X", "O"   
+
+        player_setup = PlayerSetup.new
+
+        player_setup.player_assignment(marker1, marker2, args)
+        expect(player_setup.player1).to be_a_kind_of(Human)
+        expect(player_setup.player2).to be_a_kind_of(Human)
+      end
+    end
+  end
+end
