+++ spec/player_spec.rb
@@ -2,8 +2,19 @@ module TicTacToe
   require 'spec_helper'
 
   describe Player do 
+    
+    let(:mock_console_ui) { double }
+    let(:board) { Board.new }
+    let(:validation) {Validation.new}
+
     it "returns the ordinal number and marker of a player" do
-      player = Player.new('X')
+      args = 
+        {
+          :board => board, 
+          :console_ui => mock_console_ui,
+          :validation => validation
+        }
+      player = Player.new('X', args)
       expect(player.marker).to eq('X')
     end
   end
